+++ main/Line.hpp
@@ -9,10 +9,10 @@ struct Line
 {
 	// (min x, min y)
 	Point min;
-	unsigned int length;
+	unsigned short length;
 	// whether or not _length_ is oriented horizontally
 	bool horizontal;
 
 	// shift _this_ by normalized _vector_, _scale_ times
-	void ApplyVector(Vector2D vector, unsigned int scale);
+	void ApplyVector(Vector2D vector, unsigned long scale);
 };
