+++ main/Vector2D.cpp
@@ -5,7 +5,7 @@ Vector2D::Vector2D() :
 {
 }
 
-Vector2D::Vector2D(const int _x, const int _y)
+Vector2D::Vector2D(const long _x, const long _y)
 {
 	x = _x;
 	y = _y;
